requirements.txt
Django>=5.2,<5.3
djangorestframework
mysqlclient
python-dotenv
requests